Added `olmChannels` variable and `DefaultChannelAttribute`
Updated `OlmGenerator.cs` to include a new variable `olmChannels` that is obtained from the `AnalyzerConfigOptions.GlobalOptions` dictionary. If `olmChannels` is null or empty, it is set to the value of `defaultChannel.DefaultChannel`. If not, all semicolons in it are replaced with commas. Also, added a new attribute `DefaultChannelAttribute` in the `Neon.Operator.OperatorLifecycleManager` namespace to specify the default channel for the Operator.