Update process name check in CustomResourceDefinitionGenerator
Removed the static readonly string array `allowedProcesses` from `CustomResourceDefinitionGenerator.cs` in the `Neon.Operator.Analyzers` namespace. Replaced the check for the current process name in the `allowedProcesses` array with a check for whether the process name contains any of the previously allowed process names ("VBCSCompiler", "testhost", "dotnet"). This check is now stored in the `shouldRunLive` variable. Also, updated the condition in the `if` statement to use the `shouldRunLive` variable instead of checking the `allowedProcesses` array.