`Update changelog generation and upload jobs for releases`
Modified the code to generate a changelog for both regular and prerelease versions in the `publish.yaml` file. The condition for generating a changelog has been updated to only generate if the release is not a prerelease. A new job has been added to generate a changelog for prereleases, which will only run if the release is a prerelease. The upload job has been split into two separate jobs for regular releases and prereleases, with conditions matching those for generating the changelogs. The upload jobs now use the output from their respective changelog generation jobs.