Update release and prerelease steps in GitHub Actions
This commit updates the action used for uploading the changelog and prerelease changelog from `svenstaro/upload-release-action@v2` to `tubone24/update_release@v1.0`. New steps `update release` and `update prerelease` have been added with an id of `update_release`. These steps now include an `if` condition to check if the release is a prerelease or not. The `GITHUB_TOKEN` environment variable has been added to these steps. The `repo_token` and `tag` inputs have been removed, while the `body` input remains the same, taking the output content from the `git-cliff` and `git-cliff-prerelease` steps respectively.